+++ scripts/ci/actions/format-action.mjs
@@ -3,10 +3,10 @@
  */
 // @ts-check
 import { setFailed, info } from '@actions/core'
-import { commitUnstagedChanges, getUnstagedChanges } from './_git_utils.mjs'
-import { isPR } from './_pr_utils.mjs'
-import { runCommand } from './_utils.mjs'
-import { ROOT } from './_common.mjs'
+import { commitUnstagedChanges, getUnstagedChanges } from './libs/_git_utils.mjs'
+import { isPR } from './libs/_pr_utils.mjs'
+import { runCommand } from './libs/_utils.mjs'
+import { ROOT } from './libs/_common.mjs'
 
 const HEAD = process.env.HEAD
 const BASE = process.env.BASE
@@ -57,8 +57,8 @@ if (canWrite && files.length > 0) {
       message: 'chore: format files',
     })
     info('Unstaged changes for formatting were committed.')
-
-    process.exit(0)
+    setFailed('Unstaged changes were committed.');
+    process.exit(1)
   }
   info('No unstaged changes found.')
   process.exit(0)
