+++ scripts/ci/actions/libs/_check_permission.mjs
